+++ tests/metrics/tickets/github/test.m
@@ -1,0 +1,8 @@
+function rv = test
+
+    %| pragma Justify (metric, "function_length",
+    %|                 "to be fixed in #666 and #42");
+
+    rv = 42;
+
+end